hugo: Use secondary colour for links in left navigation sidebar
I'm no SCSS expert, but by the look of it, our value for $link-color set
in _variables_project.scss may be overwritten by Docsy's value in its
_colors.scss file, included after _variables_project.scss but before
_sidebar-tree.scss (where we make the links for the navigation sidebar
use $link-color).

Let's redefine the colour explicitely in our _styles_project.scss, and
slightly improve the formatting of the custom SCSS we have.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>